scripts/requirements.txt
boto3>=1.35.0
feedparser>=6.0.11
requests>=2.31.0
beautifulsoup4>=4.12.3
PyYAML>=6.0.1
Pillow>=10.4.0
arxiv>=2.1.3
lxml>=5.3.0
python-dateutil>=2.9.0
cairosvg>=2.7.0
